requirements.txt
opencv-python == 4.12.0.88
keras == 3.10.0
keras-facenet == 0.3.2
pandas == 2.3.3
numpy == 1.26.4
tensorflow == 2.19.0
mediapipe == 0.10.21
scikit-learn == 1.7.2
openpyxl == 3.1.5
ultralytics == 8.3.213
Pillow == 11.3.0
flask == 3.1.2